added Tag component
